refactor(utils): separate formatters [#123]
