+++ utils/localization.js
@@ -1,4 +1,4 @@
-import { DEFAULT_LOCALE } from 'constants';
+import { DEFAULT_LOCALE, VALID_LOCALES } from 'constants';
 
 // here are the rules for localization:
 // 1. requested lang
@@ -22,3 +22,21 @@ export const localizeFields = (fields, format = copy) => (object, lang) => {
   });
   return localized;
 };
+
+export const getLocale = ({ asPath, query: { lang } }) => {
+  // query has 'lang' field if it was successfully matched by the router.
+  if (lang) {
+    return lang;
+  }
+
+  // parse locale for invalid paths (404)
+  // asPath starts with '/', so we have to take [1], not [0].
+  const parsedLang = asPath.split('/')[1];
+  if (VALID_LOCALES.includes(parsedLang)) {
+    return parsedLang;
+  }
+
+  return DEFAULT_LOCALE;
+};
+
+export const replaceLocale = (url, lang) => url.replace(`/${lang}`, '');
